requirements.txt
amqp==1.4.9
aniso8601==3.0.2
anyjson==0.3.3
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.3
backcall==0.1.0
billiard==3.3.0.23
blessings==1.7
bpython==0.17.1
celery==3.1.26.post2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
configparser==3.5.0
cryptography==2.3
curtsies==0.3.0
decorator==4.4.0
defusedxml==0.5.0
Django==1.11.2
django-celery==3.2.2
django-crispy-forms==1.7.2
django-formtools==2.1
django-redis==4.10.0
django-redis-sessions==0.6.1
djangorestframework==3.5.4
djangorestframework-jwt==1.11.0
dnspython==1.15.0
entrypoints==0.3
eventlet==0.24.1
flake8==3.7.7
flake8-polyfill==1.0.2
Flask==1.0.2
gevent==1.3.7
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
greenlet==0.4.15
httplib2==0.9.2
idna==2.7
importlib-metadata==0.8
ipython==6.2.0
ipython-genutils==0.2.0
isort==4.3.16
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
kombu==3.0.37
lazy-object-proxy==1.3.1
mando==0.6.4
MarkupSafe==1.1.1
mccabe==0.6.1
mock==2.0.0
mongo==0.2.0
mongoengine==0.17.0
monotonic==1.5
nameko==2.11.0
oauthlib==2.1.0
parso==0.3.4
path.py==11.5.0
pbr==5.1.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.4.1
promise==2.2.1
prompt-toolkit==1.0.16
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.18
pyflakes==2.1.1
pyftpdlib==1.5.4
pygame==1.9.4
Pygments==2.3.1
PyJWT==1.6.4
pylint==2.3.1
pymongo==3.7.1
PyMySQL==0.9.2
python3-memcached==1.51
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
radon==3.0.1
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
Rx==1.6.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
traitlets==4.3.2
typed-ast==1.3.1
typing==3.6.6
urllib3==1.23
vine==1.1.4
virtualenv==16.0.0
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.11.1
xadmin==0.6.1
zipp==0.3.3
